Organize TCP module into a package
